Close out orders when employee hits "Confirm"
This introduces a bug
where if an employee fills out `cash_handled` and clicks "Confirm"
before setting an end time,
nothing will happen.

Then, when the employee taps the "End Time" field,
the order will immediately close with the current time stored as the end
time.

This bug will be fixed up with a planned feature, #34.
That feature ensures that the end time is always set
before the employee can close the order.

Closes #40.