Rename required props away from `data`.
Insisting on calling all props `data` is a bit limiting,
especially when we can acheive the same thing (tracking props)
through diligent use of the `PropTypes` library.

We already have an example where a single prop named `data`
doesn't provide enough flexibility;
we had to pass in an `Order` and a list of `Extra`s
into the `Order` component.