Start to define standard API for data in the store
The store should be fairly agnostic of the type of data it holds,
while at the same time providing clean ways to get at and process
the data.

This is a start to the process of pulling data-specific code
out of the store and into a separate file.

There is still much code in the store that is tied to data models,
and we'll tackle it slowly as we go.

See next commit for a cleaner diff of how this changes the API.