Swap out `react-router-dom` for `aviator`
This gives us more control over the page's URL behavior;
For example, it may let us use a tree-based data structure
to determine what appears on the page, instead of a one-dimensional URL.

I expect this to become important down the road.
For now, it simply makes the behavior a bit easier to understand.