Store visible order information directly
Previously, we were string-typing this information,
as an artifact of our old URL routing logic.

This new approach is slightly more robust.

This commit also removes a number of props
from components throughout the app.
This change indicates a preference for accessing data and functions
directly from the near-universal `assembly` prop.